requirements.txt
APScheduler==3.8.1
backports.zoneinfo==0.2.1
black==21.12b0
click==8.0.3
Flask==2.0.2
Flask-APScheduler==1.12.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
speedtest-cli==2.1.3
tomli==1.2.3
typed-ast==1.5.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
Werkzeug==2.0.2
zipp==3.7.0
